Modifications to performance scripts
- Support for NERSC cori (SLURM scheduler)
- Add BENCH_FLAGS env variable to specify compile time arguments to benchmarks
- Add WORKLOAD_ARGS env variable to specify runtime arguments to benchmarks
- Add OCRRUN_OPT_TPL_NODEFILE variable to select nodes from a nodefile template
- Add duration to generated report in addition to throughput
- Changed timer's nb instances to unsigned long long

+++ ocr/tests/performance-tests/platform/x86.c
@@ -21,9 +21,9 @@ double elapsed_sec(struct timeval * t_start, struct timeval * t_stop) {
     return (((double)elapsed)/1000000);
 }
 
-void print_throughput(char * timer_name, long nb_instances, double time_sec) {
+void print_throughput(char * timer_name, unsigned long long nb_instances, double time_sec) {
     printf("Timer Name        : %s\n", timer_name);
-    printf("Workload    (unit): %ld\n", nb_instances);
+    printf("Workload    (unit): %llu\n", nb_instances);
     printf("Duration    (s)   : %f\n", time_sec);
     printf("Throughput  (op/s): %f\n", nb_instances/time_sec);
 }
@@ -36,12 +36,12 @@ void print_elapsed_usec(long time_usec) {
     printf("Elapsed       (us): %ld\n", time_usec);
 }
 
-void summary_throughput_timer(timestamp_t * start_t, timestamp_t * stop_t, int instances) {
+void summary_throughput_timer(timestamp_t * start_t, timestamp_t * stop_t, unsigned long long instances) {
     double secs = elapsed_sec(start_t, stop_t);
     summary_throughput_dbl(secs, instances);
 }
 
-void summary_throughput_dbl(double secs, int instances) {
+void summary_throughput_dbl(double secs, unsigned long long instances) {
     print_elapsed(secs);
     print_throughput("Summary", instances, secs);
 }
